+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - feat: `execute` command which runs shell commands with the environment variables loaded from .env* files
 - fix: exit app if special files could not be loaded
 - feat: `run` command without scripts will run `go run .` now
-- feat: `postbuild`, `postinstall`, `prebuild` and `preinstall` script support
+- feat: `postbuild`, `postinstall`, `posttest`, `prebuild`, `preinstall` and `pretest` script support
 - feat: add `--no-script` flags for `build`, `start`, `test` and `tidy` 
 
 ## 0.10.1